requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
bcrypt
python-dotenv
pyjwt